+++ packages/server/api/src/app/flows/flow.module.ts
@@ -24,7 +24,6 @@ import { flowVersionService } from './flow-version/flow-version.service';
 import { flowVersionController } from './flow/flow-version.controller';
 import { flowController } from './flow/flow.controller';
 import { stepRunService } from './step-run/step-run-service';
-import { testTriggerController } from './test-trigger/test-trigger-controller';
 import { testController } from './test/test.controller';
 
 export const flowModule: FastifyPluginAsyncTypebox = async (app) => {
@@ -32,7 +31,6 @@ export const flowModule: FastifyPluginAsyncTypebox = async (app) => {
   await app.register(flowVersionController, { prefix: '/v1/flow-versions' });
   await app.register(flowController, { prefix: '/v1/flows' });
   await app.register(testController, { prefix: '/v1/test' });
-  await app.register(testTriggerController, { prefix: '/v1/test-trigger' });
   websocketService.addListener(WebsocketServerEvent.TEST_FLOW_RUN, (socket) => {
     return async (data: TestFlowRunRequestBody) => {
       let principal;
